Crate solidity_abi

Structs

ABIDecodeError

ABIListWriter

LinkedBytes

Traits

ABIDecodable

A type implements ABIDecodable must be a tuple of types implement ABIVariable. For convenient, for the tuple with only one element like (U256,), we implement ABIDecodable for U256 instead.

ABIEncodable

A type implements ABIEncodable must be a tuple of types implement ABIVariable. For convenient, for the tuple with only one element like (U256,), we implement ABIEncodable for U256 instead.

ABIPackedEncodable

A type implements ABIPackedEncodable must be a tuple of types implement ABIVariable. For convenient, for the tuple with only one element like (U256,), we implement ABIPackedEncodable for U256 instead.

ABIVariable

EventIndexEncodable

A type implements EventIndexEncodable must be a tuple of types implement EventIndexEncodable. For convenient, for the tuple with only one element like (U256,), we implement EventIndexEncodable for U256 instead.

Functions

read_abi_list

string_revert_reason_decode