Trait solidity_abi::ABIDecodable

pub trait ABIDecodable: Sized {
    // Required method
    fn abi_decode(data: &[u8]) -> Result<Self, ABIDecodeError>;
}

A type implements ABIDecodable must be a tuple of types implement ABIVariable. For convenient, for the tuple with only one element like (U256,), we implement ABIDecodable for U256 instead.

Required Methods

fn abi_decode(data: &[u8]) -> Result<Self, ABIDecodeError>

Object Safety

This trait is not object safe.