Trait solidity_abi::ABIPackedEncodable

pub trait ABIPackedEncodable: Sized {
    // Required method
    fn abi_packed_encode(&self) -> Vec<u8> ;
}

A type implements ABIPackedEncodable must be a tuple of types implement ABIVariable. For convenient, for the tuple with only one element like (U256,), we implement ABIPackedEncodable for U256 instead.

Required Methods

fn abi_packed_encode(&self) -> Vec<u8>

Object Safety

This trait is not object safe.